break-out Bar, Variance, and MarkedIndicators
